requirements.txt
langchain
langchain-core
langchain-community
langchain-experimental
langchain-openai
langgraph
langserve[all]
langchain-cli
langsmith
langchain-groq
groq
yfinance>=0.2.36
python-dotenv
fastapi>=0.109.0
uvicorn>=0.27.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas>=2.0.0
lxml==4.9.3
selenium==4.16.0
aiohttp==3.9.1
tqdm==4.66.1
pytrends>=4.9.0
requests_html
numpy>=1.24.0
matplotlib>=3.7.0
yahoo_fin>=0.8.9
Jinja2>=3.1.2
